# Precompile some functions for faster TTFX