--- a/src/precompile.jl
+++ b/src/precompile.jl
@@ -1,0 +1,10 @@
+
+using PrecompileTools
+
+@setup_workload let
+    @compile_workload begin
+        setmacro(identity, :(a.x = 1), overwrite=false)
+        setmacro(identity, :(a.x = 1), overwrite=true)
+        insertmacro(identity, :(a.x = 1), overwrite=false)
+    end
+end