This has not a big impact but it works, e.g. now

```julia
julia> using Accessors

julia> struct Q k::Float64 end

julia> q = Q(1.0)
Q(1.0)

julia> @time @eval f(q) = @set q.k = 2.0;
  0.007665 seconds (2.08 k allocations: 161.078 KiB, 91.12% compilation time) # this pr
  0.021409 seconds (14.66 k allocations: 1023.523 KiB, 96.41% compilation time) # master

julia> @time @eval f(q);
  0.032977 seconds (23.63 k allocations: 1.616 MiB, 99.53% compilation time) # this pr
  0.032714 seconds (23.67 k allocations: 1.628 MiB, 99.52% compilation time) # master
```

Addresses partially #173