@Tortar 2024-09-22
I guess that the time spent on function instantiation is anyway happening at precompile time inside packages though.