@jw3126 2024-09-26
Mhh the gain is minor, I wonder if it is worth the additional dependency. How does this impact precompile time?