@Tortar 2024-09-26
almost the same for precompile times (maybe 0.1 second more), but actually the additional dependency impacts a bit the importing:

```julia
julia> @time using Accessors
  0.039866 seconds (36.46 k allocations: 4.357 MiB, 10.73% compilation time) # this pr
  0.031610 seconds (25.88 k allocations: 3.539 MiB, 13.73% compilation time) # master

julia> struct Q k::Float64 end

julia> q = Q(1.0)
Q(1.0)

julia> @time @eval f(q) = @set q.k = 2.0;
  0.008145 seconds (2.08 k allocations: 144.578 KiB, 88.98% compilation time) # this pr
  0.022843 seconds (14.67 k allocations: 1.015 MiB, 95.64% compilation time) # master

julia> @time f(q);
  0.034677 seconds (23.59 k allocations: 1.614 MiB, 99.95% compilation time) # this pr
  0.034935 seconds (23.64 k allocations: 1.632 MiB, 99.93% compilation time) # master
```

so I close for now, let me know if you think it's still worth it