@MasonProtter 2024-12-02
One possibility would be to add a weak-dep / extension targetting PrecompileTools.jl, so that if another package is already going to bring in PrecompileTools, then we might as well also do some precompilation, but if not, we can skip it.